+++ projects/swimlane/ngx-ui/src/lib/components/radiobutton/radiobutton-group.component.ts
@@ -11,14 +11,17 @@ import {
   AfterContentInit,
   ChangeDetectionStrategy,
   ChangeDetectorRef,
-  OnChanges
+  OnChanges,
+  HostListener,
+  HostBinding
 } from '@angular/core';
 import { NG_VALUE_ACCESSOR, ControlValueAccessor } from '@angular/forms';
-import { coerceBooleanProperty } from '@angular/cdk/coercion';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 
 import { RadioButtonComponent } from './radiobutton.component';
+import { KeyboardKeys } from '../../enums/keyboard-keys.enum';
+import { coerceBooleanProperty, coerceNumberProperty } from '@angular/cdk/coercion';
 
 const RADIOGROUP_VALUE_ACCESSOR = {
   provide: NG_VALUE_ACCESSOR,
@@ -27,6 +30,10 @@ const RADIOGROUP_VALUE_ACCESSOR = {
 };
 
 let nextId = 0;
+
+function mod(v: number, n: number): number {
+  return ((v % n) + n) % n;
+}
 
 @Component({
   exportAs: 'ngxRadiobuttonGroup',
@@ -50,8 +57,9 @@ export class RadioButtonGroupComponent implements ControlValueAccessor, OnDestro
   get disabled() {
     return this._disabled;
   }
-  set disabled(disabled: boolean) {
-    this._disabled = coerceBooleanProperty(disabled);
+  set disabled(val: boolean) {
+    this._disabled = coerceBooleanProperty(val);
+    this._updateRadioDisabledState();
   }
 
   @Input()
@@ -61,8 +69,7 @@ export class RadioButtonGroupComponent implements ControlValueAccessor, OnDestro
   set value(value) {
     if (this._value !== value) {
       this._value = value;
-      this._updateSelectedRadioFromValue();
-      this._updateRadioDisabledState();
+      this.update();
       this.onChangeCallback(this._value);
     }
   }
@@ -78,6 +85,24 @@ export class RadioButtonGroupComponent implements ControlValueAccessor, OnDestro
     }
   }
 
+  @Input()
+  get focusIndex() {
+    return this._focusIndex;
+  }
+  set focusIndex(val: number) {
+    this._focusIndex = coerceNumberProperty(val);
+    this.focusOn(this._focusIndex);
+  }
+
+  @HostBinding('attr.tabindex')
+  @Input()
+  get tabIndex() {
+    return this.disabled ? -1 : this._tabIndex;
+  }
+  set tabIndex(val: number) {
+    this._tabIndex = coerceNumberProperty(val);
+  }
+
   @Output() change = new EventEmitter<boolean>();
   @Output() blur = new EventEmitter<Event>();
   @Output() focus = new EventEmitter<FocusEvent>();
@@ -91,8 +116,10 @@ export class RadioButtonGroupComponent implements ControlValueAccessor, OnDestro
 
   private _name: string = this.UNIQUE_ID;
   private _value = false;
-  private _selected: RadioButtonComponent;
   private _disabled = false;
+  private _selected: RadioButtonComponent;
+  private _focusIndex = -1;
+  private _tabIndex = 0;
   private _destroy$ = new Subject<void>();
 
   constructor(private readonly _cdr: ChangeDetectorRef) {}
@@ -104,6 +131,8 @@ export class RadioButtonGroupComponent implements ControlValueAccessor, OnDestro
     if (this._radios) {
       this._radios.changes.subscribe(this.subscribeToRadios.bind(this));
     }
+
+    this.update();
   }
 
   ngOnDestroy() {
@@ -112,7 +141,38 @@ export class RadioButtonGroupComponent implements ControlValueAccessor, OnDestro
   }
 
   ngOnChanges() {
-    this._updateRadioDisabledState();
+    this.update();
+  }
+
+  @HostListener('focus')
+  onFocus() {
+    if (this.selected) {
+      // Moves keyboard focus to the checked radio button in a radiogroup.
+      this.focusIndex = this._radios.toArray().indexOf(this.selected);
+    } else {
+      // If no radio button is checked, focus moves to the first radio button in the group.
+      this.focusFirst();
+    }
+  }
+
+  @HostListener('keydown', ['$event'])
+  onKeyUp(ev: KeyboardEvent) {
+    switch (ev.code) {
+      case KeyboardKeys.ARROW_LEFT:
+      case KeyboardKeys.ARROW_UP:
+        ev.stopPropagation();
+        ev.preventDefault();
+        this.focusIn(-1); // Moves focus to previous radio button in the group.
+        this.selectIndex(this.focusIndex); // Selects the radio button in the group.
+        break;
+      case KeyboardKeys.ARROW_RIGHT:
+      case KeyboardKeys.ARROW_DOWN:
+        ev.stopPropagation();
+        ev.preventDefault();
+        this.focusIn(1); // Moves focus to next radio button in the group.
+        this.selectIndex(this.focusIndex); // Selects the radio button in the group.
+        break;
+    }
   }
 
   subscribeToRadios(): void {
@@ -125,7 +185,7 @@ export class RadioButtonGroupComponent implements ControlValueAccessor, OnDestro
       });
     }
 
-    this._cdr.markForCheck();
+    this.update();
   }
 
   onRadioSelected(value: string) {
@@ -138,6 +198,7 @@ export class RadioButtonGroupComponent implements ControlValueAccessor, OnDestro
 
   writeValue(value: any): void {
     this.value = value;
+    this.update();
   }
 
   registerOnChange(fn: any): void {
@@ -157,6 +218,57 @@ export class RadioButtonGroupComponent implements ControlValueAccessor, OnDestro
     // placeholder
   }
 
+  focusFirst() {
+    if (!this.disabled && this._radios) {
+      const len = this._radios.length;
+      for (let i = 0; i < len; i++) {
+        if (!this._radios.get(i).disabled) {
+          this.focusIndex = i;
+          break;
+        }
+      }
+    }
+  }
+
+  setDisabledState(isDisabled: boolean): void {
+    this.disabled = coerceBooleanProperty(isDisabled);
+  }
+
+  private selectIndex(index: number) {
+    if (!this.disabled && this.focusIndex > -1) {
+      this.value = this._radios.get(index).value;
+    }
+  }
+
+  private focusOn(index: number) {
+    if (!this.disabled) {
+      this._radios.get(index).focusElement();
+    }
+  }
+
+  /**
+   * Moves focus to next radio button in the group.
+   * +1 is next radio button, -1 is previous radio button.
+   */
+  private focusIn(dir: 1 | -1) {
+    if (!this.disabled && this._radios) {
+      const len = this._radios.length;
+      for (let i = 1; i < len; i++) {
+        const ii = mod(this.focusIndex + dir * i, len);
+        if (!this._radios.get(ii).disabled) {
+          this.focusIndex = ii;
+          return;
+        }
+      }
+    }
+  }
+
+  private update() {
+    this._updateSelectedRadioFromValue();
+    this._updateRadioDisabledState();
+    this._cdr.markForCheck();
+  }
+
   private _updateRadioButtonNames(): void {
     if (this._radios) {
       this._radios.forEach(radio => {
@@ -170,6 +282,7 @@ export class RadioButtonGroupComponent implements ControlValueAccessor, OnDestro
     if (this._radios) {
       this._radios.forEach(radio => {
         radio.checked = this.value === radio.value;
+        radio.isInGroup = true;
 
         if (radio.checked) {
           this._selected = radio;