+++ projects/swimlane/ngx-ui/src/lib/components/radiobutton/radiobutton.component.ts
@@ -6,10 +6,12 @@ import {
   ViewEncapsulation,
   forwardRef,
   ChangeDetectionStrategy,
-  ChangeDetectorRef
+  ChangeDetectorRef,
+  ElementRef
 } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
-import { coerceBooleanProperty, coerceNumberProperty } from '@angular/cdk/coercion';
+import { coerceBooleanProperty } from '@angular/cdk/coercion';
+import { CoerceNumberProperty } from '../../utils/coerce/coerce-number';
 
 const RADIO_VALUE_ACCESSOR = {
   provide: NG_VALUE_ACCESSOR,
@@ -38,13 +40,11 @@ export class RadioButtonComponent implements ControlValueAccessor {
   @Input() id: string = this.UNIQUE_ID;
   @Input() name: string = this.UNIQUE_ID;
 
+  @Input() radioId = `${this.id}-radio`;
+
   @Input()
-  get tabindex() {
-    return this._tabindex;
-  }
-  set tabindex(tabindex: number) {
-    this._tabindex = coerceNumberProperty(tabindex);
-  }
+  @CoerceNumberProperty()
+  tabindex = 0;
 
   @Input()
   get checked() {
@@ -58,14 +58,15 @@ export class RadioButtonComponent implements ControlValueAccessor {
       this.onChangeCallback(this._value);
     }
 
-    this.cdr.markForCheck();
+    this.cdr.detectChanges(); // Update this component
+    this.cdr.markForCheck(); // Update the host component (radio group)
   }
 
   @Input()
-  get value() {
+  get value(): any {
     return this._value;
   }
-  set value(value: boolean) {
+  set value(value: any) {
     if (this._value !== value) {
       this._value = value;
       this.onChangeCallback(this._value);
@@ -80,18 +81,18 @@ export class RadioButtonComponent implements ControlValueAccessor {
     this._disabled = coerceBooleanProperty(disabled);
   }
 
-  @Output() change = new EventEmitter<boolean>();
+  @Output() change = new EventEmitter<any>();
   @Output() blur = new EventEmitter<Event>();
   @Output() focus = new EventEmitter<FocusEvent>();
 
   public groupDisabled = false;
+  public isInGroup = false;
 
   private _checked = false;
   private _value = false;
   private _disabled = false;
-  private _tabindex = 0;
 
-  constructor(private readonly cdr: ChangeDetectorRef) {}
+  constructor(private readonly cdr: ChangeDetectorRef, private readonly elementRef: ElementRef) {}
 
   _onInputChange(event: Event) {
     event.stopPropagation();
@@ -107,6 +108,10 @@ export class RadioButtonComponent implements ControlValueAccessor {
     this.onTouchedCallback();
   }
 
+  focusElement(): void {
+    this.elementRef.nativeElement.getElementsByClassName('radio-label')[0].focus();
+  }
+
   registerOnChange(fn: any): void {
     this.onChangeCallback = fn;
   }
@@ -115,7 +120,14 @@ export class RadioButtonComponent implements ControlValueAccessor {
     this.onTouchedCallback = fn;
   }
 
-  private onChangeCallback(value: boolean) {
+  onSpace(ev: Event) {
+    // If the radio button with focus is unchecked, it's state will be changed to checked.
+    ev.stopPropagation();
+    ev.preventDefault();
+    this.checked = true;
+  }
+
+  private onChangeCallback(value: any) {
     if (this.checked) {
       this.change.emit(value);
     }